mcp-db/pyproject.toml
[project]
name = "mcp-db"
version = "0.1.0"
description = "MCP stdio server for read-only DB access (Cursor + backend DB)"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0,<2",
    "sqlalchemy>=2.0,<3",
    "psycopg[binary]>=3.2,<4",
    "python-dotenv>=1.0,<2",
]

[tool.uv]
package = false
